OpenSolaris knows (u)int64_t nowadays.
Also the standard int64_t type is typedefed as
"long int" with OpenSolaris 64bit gcc making it
incompatible with a "long long int" typedefed
int64.

git-svn-id: https://matroska.svn.sourceforge.net/svnroot/matroska/trunk/libebml@48 a6f86f6d-0131-4f8e-9e7b-e335508773d5

+++ ebml/c/libebml_t.h
@@ -89,12 +89,12 @@ extern "C" {
 # ifdef _NO_LONGLONG
 #  error This compiler does not support 64bit integers.
 # endif
-    typedef long long int64; // int64_t is not always defined :(
+    typedef int64_t int64;
     typedef int32_t int32;
     typedef int16_t int16;
     typedef int8_t int8;
     typedef int8_t character;
-    typedef unsigned long long uint64; // uint64_t is not always defined :(
+    typedef uint64_t uint64;
     typedef uint32_t uint32;
     typedef uint16_t uint16;
     typedef uint8_t uint8;
